Refactor mappers into one service
